Check positioner motion when constructing TargetsAvailable.
The TargetsAvailable class stores the initial list of targets
to consider for each positioner.  This is before any checks for
collisions or the remaining obs for each target.  Previously,
the check for whether these targets are physically reachable was
done at the same time as collision checks during assignment.
This work moves that "physically reachable" check to the earlier
step when computing TargetsAvailable.  So now the available targets
that are written to the output include this constraint.

Closes #263.