Try to remove files before overwriting them.
It doesn't seem like this should be necessary.  But at KPNO, when originally constructing files
as the planner user, and then constructing them again as the msdos user, even if the msdos user
has write permission to the planner-created file, this can lead to exceptions.  But they go
away if the files are removed before overwriting.  So that's the approach taken here.