Added behavioral switch based on FIBERASSIGN_BEHAVIOR env var: If 0 -> use old 'buggy' compiler specific ::abs call If 1 or undefined, use ::fabs
While implementing this I found an interesting behavior that should be
documented and took me a while to figure out.  If I include <stdlib.h>
then ::abs always performs correctly as a floating point abs.  I had
thought I needed to include stdlib.h to handle the env var, though it
turned out I did not.

This may be useful in tracking down exactly where the discrepancy occurs
since we see a difference between desi 23.10 / gcc 11.2 and desi 24.4 /
gcc 11.2 - it is probably in libstdc++ or glibc or some other library
module that is loaded.